requirements.txt
click
pyopencl
base58
PyNaCl
